PYTHON-6013 Consolidate session/transaction helpers into a shared file
